refactor: 외부 api 호출 하는 부분 트랜잭션 외부로 분리
